[Chore] ZoocGradientButton에 style함수 추가 #134
