+++ ZOOC/ZOOC/Presentation/Onboarding/View/OnboardingAgreementView.swift
@@ -66,9 +66,6 @@ final class OnboardingAgreementView: OnboardingBaseView {
         
         signUpButton.do {
             $0.setTitle("회원가입", for: .normal)
-            $0.setTitleColor(.zoocWhite1, for: .normal)
-            $0.titleLabel?.font = .zoocSubhead1
-            $0.titleLabel?.textAlignment = .center
             $0.isEnabled = false
         }
     }
