+++ ZOOC/ZOOC/Presentation/Onboarding/View/OnboardingChooseRoleView.swift
@@ -71,9 +71,6 @@ final class OnboardingChooseRoleView: OnboardingBaseView {
         
         nextButton.do {
             $0.setTitle("이렇게 불러주세요", for: .normal)
-            $0.setTitleColor(.white, for: .normal)
-            $0.titleLabel?.font = .zoocSubhead1
-            $0.titleLabel?.textAlignment = .center
             $0.isEnabled = false
         }
     }
