+++ ZOOC/ZOOC/Presentation/Onboarding/View/OnboardingInviteFamilyView.swift
@@ -67,13 +67,10 @@ final class OnboardingReInviteFamilyView: OnboardingBaseView {
             $0.setTitleColor(.zoocGray1, for: .normal)
             $0.titleLabel?.font = .zoocSubhead1
             $0.setUnderline()
-            
         }
         
         inviteButton.do {
             $0.setTitle("다음", for: .normal)
-            $0.setTitleColor(.zoocWhite1, for: .normal)
-            $0.titleLabel?.font = .zoocSubhead1
         }
     }
